@@ -1,12 +1,24 @@
 //	 ,+---+
 //	+---+´|    HASHBOX SOURCE
-//	| # | |    Copyright 2015-2024
+//	| # | |    Copyright 2015-2026
 //	+---+´
 
 package main
 
 import (
 	//"github.com/davecheney/profile"
+	"bytes"
+	"crypto/rand"
+	"encoding/binary"
+	"fmt"
+	"net"
+	"os"
+	"os/signal"
+	"path/filepath"
+	"strconv"
+	"syscall"
+	"time"
+
 	"net/http"
 	_ "net/http/pprof"
 	"runtime"
@@ -14,27 +26,15 @@ import (
 
 	"github.com/fredli74/bytearray"
 	cmd "github.com/fredli74/cmdparser"
-	"github.com/fredli74/hashbox/core"
+	"github.com/fredli74/hashbox/pkg/accountdb"
+	"github.com/fredli74/hashbox/pkg/core"
+	"github.com/fredli74/hashbox/pkg/storagedb"
 	"github.com/fredli74/lockfile"
-
-	"bytes"
-	"encoding/binary"
-	"fmt"
-	"math/rand"
-	"net"
-	"os"
-	"os/signal"
-	"path/filepath"
-	"strconv"
-	"sync"
-	"time"
-
-	"github.com/kardianos/osext"
 )
 
 var Version = "(dev-build)"
 
-const DEFAULT_SERVER_IP_PORT int = 7411
+const DEFAULT_SERVER_IP_PORT int = core.DefaultServerPort
 
 var datDirectory string
 var idxDirectory string
@@ -42,35 +42,6 @@ var accountHandler *AccountHandler
 var storageHandler *StorageHandler
 
 var done chan bool
-
-var logLock sync.Mutex
-
-// ASSERTS should be disabled on release
-// ASSERTS are not error handling
-// ASSERTS are tests of development errors
-func ASSERT(ok bool, v ...interface{}) {
-	if !ok {
-		_, file, line, _ := runtime.Caller(1)
-		panic(fmt.Errorf("ASSERT raised by line %v:%v %v", file, line, v))
-	}
-}
-func abort(format string, a ...interface{}) {
-	panic(fmt.Errorf(format, a...))
-}
-func abortOn(err error) {
-	if err != nil {
-		panic(err)
-	}
-}
-
-/* THIS SHOULD BE REMOVED, WE SHOULD NEVER USE PANICS AS A NORMAL EXECUTION STATE */
-func Try(f func()) (err interface{}) {
-	defer func() {
-		err = recover()
-	}()
-	f()
-	return nil
-}
 
 func handleConnection(conn net.Conn) {
 	remoteID := conn.RemoteAddr().String()
@@ -83,14 +54,18 @@ func handleConnection(conn net.Conn) {
 		}
 
 		core.Log(core.LogInfo, "%s - Connection closed", remoteID)
-		conn.Close()
+		if err := conn.Close(); err != nil {
+			core.Log(core.LogWarning, "%s - Error closing connection: %v", remoteID, err)
+		}
 	}()
 
 	var clientSession core.Session
 	var sessionAuthenticated = false
 
 	for keepAlive := true; keepAlive; {
-		conn.SetReadDeadline(time.Now().Add(10 * time.Minute)) // max size 512kb over 10 minutes is slower than 9600bps so we should be safe
+		if err := conn.SetReadDeadline(time.Now().Add(10 * time.Minute)); err != nil {
+			core.AbortOnError(err)
+		} // max size 512kb over 10 minutes is slower than 9600bps so we should be safe
 		// TODO: testing non-idle of 10s between commands
 
 		unauthorized := false
@@ -117,8 +92,8 @@ func handleConnection(conn net.Conn) {
 				} else {
 					// Create server nonce using  64-bit time and 64-bit random
 					binary.BigEndian.PutUint64(clientSession.SessionNonce[0:], uint64(time.Now().UnixNano()))
-					binary.BigEndian.PutUint32(clientSession.SessionNonce[8:], rand.Uint32())
-					binary.BigEndian.PutUint32(clientSession.SessionNonce[12:], rand.Uint32())
+					_, err := rand.Read(clientSession.SessionNonce[8:])
+					core.AbortOnError(err)
 					reply.Data = &core.MsgServerGreeting{SessionNonce: clientSession.SessionNonce}
 				}
 			case core.MsgTypeAuthenticate:
@@ -143,9 +118,9 @@ func handleConnection(conn net.Conn) {
 						if !unauthorized {
 							if !storageHandler.doesBlockExist(c.State.BlockID) {
 								reply.Type = core.MsgTypeError & core.MsgTypeServerMask
-								reply.Data = &core.MsgServerError{ErrorMessage: "Dataset pointing to a non existent block"}
+								reply.Data = &core.MsgServerError{ErrorMessage: "Dataset pointing to a nonexistent block"}
 							} else {
-								accountHandler.AddDatasetState(c.AccountNameH, c.DatasetName, c.State)
+								core.AbortOnError(accountHandler.AddDatasetState(c.AccountNameH, c.DatasetName, c.State))
 								// No need to set any data in reply
 							}
 						}
@@ -153,7 +128,7 @@ func handleConnection(conn net.Conn) {
 						c := incoming.Data.(*core.MsgClientRemoveDatasetState)
 						unauthorized = c.AccountNameH != clientSession.AccountNameH // TODO: admin support for other accounts hashes?
 						if !unauthorized {
-							accountHandler.RemoveDatasetState(c.AccountNameH, c.DatasetName, c.StateID)
+							core.AbortOnError(accountHandler.RemoveDatasetState(c.AccountNameH, c.DatasetName, c.StateID))
 							// No need to set any data in reply
 						}
 					case core.MsgTypeListDataset:
@@ -162,7 +137,7 @@ func handleConnection(conn net.Conn) {
 						if !unauthorized {
 							list := accountHandler.ListDataset(c.AccountNameH, c.DatasetName)
 							if list == nil {
-								list = new(dbStateCollection)
+								list = new(accountdb.DBStateCollection)
 							}
 							reply.Data = &core.MsgServerListDataset{States: list.States, ListH: list.ListH}
 							//				} else {
@@ -208,11 +183,11 @@ func handleConnection(conn net.Conn) {
 							for _, l := range c.Block.Links {
 								if !storageHandler.doesBlockExist(l) {
 									reply.Type = core.MsgTypeError & core.MsgTypeServerMask
-									reply.Data = &core.MsgServerError{ErrorMessage: "Linked to non existent block"}
+									reply.Data = &core.MsgServerError{ErrorMessage: core.String(fmt.Sprintf("Linked to nonexistent block: block=%x link=%x", c.Block.BlockID, l))}
 									break
 								}
 							}
-							if !storageHandler.checkFree(int64(c.Block.Data.Len())) {
+							if !storageHandler.store.CheckFree(int64(c.Block.Data.Len())) {
 								reply.Type = core.MsgTypeError & core.MsgTypeServerMask
 								reply.Data = &core.MsgServerError{ErrorMessage: "Write permission is denied because the server is out of space"}
 							}
@@ -226,7 +201,7 @@ func handleConnection(conn net.Conn) {
 							reply.Data = &core.MsgServerError{ErrorMessage: "Unable to verify blockID"}
 						}
 					default:
-						ASSERT(false, incoming.String()) // Well if we reach this point, we have not implemented everything correctly
+						core.ASSERT(false, incoming.String()) // Well if we reach this point, we have not implemented everything correctly
 					}
 				}
 			}
@@ -261,13 +236,16 @@ func run() (returnValue int) {
 
 	var err error
 
-	exeRoot, _ := osext.ExecutableFolder()
+	core.ApplyEnvUMASK()
 
 	var serverPort int64 = int64(DEFAULT_SERVER_IP_PORT)
-	datDirectory = filepath.Join(exeRoot, "data")
-	idxDirectory = filepath.Join(exeRoot, "index")
+	datDirectory, err = filepath.Abs("data")
+	core.AbortOnError(err, "abs data dir: %v", err)
+	idxDirectory, err = filepath.Abs("index")
+	core.AbortOnError(err, "abs index dir: %v", err)
 
 	cmd.Title = fmt.Sprintf("Hashbox Server %s", Version)
+	cmd.ShowCurrentDefaults = true
 	cmd.IntOption("port", "", "<port>", "Server listening port", &serverPort, cmd.Standard)
 	cmd.StringOption("data", "", "<path>", "Full path to dat files", &datDirectory, cmd.Standard)
 	cmd.StringOption("index", "", "<path>", "Full path to idx and meta files", &idxDirectory, cmd.Standard)
@@ -283,26 +261,23 @@ func run() (returnValue int) {
 		}
 	})
 
-	// Please note that datPath has not been set until we have parsed arguments, that is ok because neither of the handlers
-	// start opening files on their own
-	// TODO: remove datPath global and send them into handlers on creation instead
-	accountHandler = NewAccountHandler()
-	defer accountHandler.Close()
-	storageHandler = NewStorageHandler()
-	defer storageHandler.Close()
-
 	cmd.Command("", "", func() { // Default
 		serverAddr := net.TCPAddr{IP: nil, Port: int(serverPort), Zone: ""}
 
 		if lock, err := lockfile.Lock(filepath.Join(datDirectory, "hashbox.lck")); err != nil {
-			abort("%v", err)
+			core.Abort("%v", err)
 		} else {
 			defer lock.Unlock()
 		}
 
+		accountHandler = NewAccountHandler()
+		defer accountHandler.Close()
+		storageHandler = NewStorageHandler()
+		defer storageHandler.Close()
+
 		var listener *net.TCPListener
 		if listener, err = net.ListenTCP("tcp", &serverAddr); err != nil {
-			abort("Error listening: %v", err.Error())
+			core.Abort("Error listening: %v", err.Error())
 		}
 		core.Log(core.LogInfo, "%s is listening on %s", cmd.Title, listener.Addr().String())
 
@@ -311,12 +286,11 @@ func run() (returnValue int) {
 
 		signalchan := make(chan os.Signal, 1)
 		defer close(signalchan)
-		signal.Notify(signalchan, os.Interrupt)
-		signal.Notify(signalchan, os.Kill)
+		signal.Notify(signalchan, os.Interrupt, syscall.SIGTERM)
 		go func() {
 			for s := range signalchan {
 				core.Log(core.LogInfo, "Received OS signal: %v", s)
-				listener.Close()
+				core.AbortOnError(listener.Close())
 				// done <- true
 				return
 			}
@@ -350,17 +324,22 @@ func run() (returnValue int) {
 	// It should be an interface to an adminsitrative tool instead
 	cmd.Command("adduser", "<username> <password>", func() {
 		if lock, err := lockfile.Lock(filepath.Join(datDirectory, "hashbox.lck")); err != nil {
-			abort("%v", err)
+			core.Abort("%v", err)
 		} else {
 			defer lock.Unlock()
 		}
 
+		accountHandler = NewAccountHandler()
+		defer accountHandler.Close()
+		storageHandler = NewStorageHandler()
+		defer storageHandler.Close()
+
 		if len(cmd.Args) < 4 {
-			abort("Missing argument to adduser command")
+			core.Abort("Missing argument to adduser command")
 		}
 
-		if (!accountHandler.SetInfo(AccountInfo{AccountName: core.String(cmd.Args[2]), AccessKey: core.GenerateAccessKey(cmd.Args[2], cmd.Args[3])})) {
-			abort("Error creating account")
+		if !accountHandler.SetInfo(accountdb.AccountInfo{AccountName: core.String(cmd.Args[2]), AccessKey: core.GenerateAccessKey(cmd.Args[2], cmd.Args[3])}) {
+			core.Abort("Error creating account")
 		}
 		accountNameH := core.Hash([]byte(cmd.Args[2]))
 		dataEncryptionKey := core.GenerateDataEncryptionKey()
@@ -368,13 +347,14 @@ func run() (returnValue int) {
 		core.EncryptDataInPlace(dataEncryptionKey[:], core.GenerateBackupKey(cmd.Args[2], cmd.Args[3]))
 
 		var blockData bytearray.ByteArray
-		blockData.Write(dataEncryptionKey[:])
+		if _, err := blockData.Write(dataEncryptionKey[:]); err != nil {
+			core.AbortOnError(err)
+		}
 		block := core.NewHashboxBlock(core.BlockDataTypeRaw, blockData, nil)
 		if !storageHandler.writeBlock(block) {
-			abort("Error writing key block")
+			core.Abort("Error writing key block")
 		}
-		err := accountHandler.AddDatasetState(accountNameH, core.String("\x07HASHBACK_DEK"), core.DatasetState{BlockID: block.BlockID})
-		abortOn(err)
+		core.AbortOnError(accountHandler.AddDatasetState(accountNameH, core.String("\x07HASHBACK_DEK"), core.DatasetState{BlockID: block.BlockID}))
 
 		block.Release()
 		core.Log(core.LogInfo, "User added")
@@ -390,10 +370,15 @@ func run() (returnValue int) {
 	cmd.IntOption("threshold", "gc", "<percentage>", "Compact minimum dead space threshold", &optGcDeadSkip, cmd.Standard)
 	cmd.Command("gc", "", func() {
 		if lock, err := lockfile.Lock(filepath.Join(datDirectory, "hashbox.lck")); err != nil {
-			abort("%v", err)
+			core.Abort("%v", err)
 		} else {
 			defer lock.Unlock()
 		}
+
+		accountHandler = NewAccountHandler()
+		defer accountHandler.Close()
+		storageHandler = NewStorageHandler()
+		defer storageHandler.Close()
 
 		start := time.Now()
 		if !optGcCompactOnly {
@@ -402,52 +387,57 @@ func run() (returnValue int) {
 			for _, r := range accountHandler.CollectAllRootBlocks(optGcForce) {
 				roots = append(roots, r.BlockID)
 			}
-			storageHandler.MarkIndexes(roots, true)
-			storageHandler.SweepIndexes(true)
+			storageHandler.store.MarkIndexes(roots, true)
+			storageHandler.store.SweepIndexes(true)
 			core.Log(core.LogInfo, "Mark and sweep duration %.1f minutes", time.Since(start).Minutes())
-			storageHandler.ShowStorageDeadSpace()
+			storageHandler.store.ShowStorageDeadSpace()
 		}
 		if optGcCompact || optGcCompactOnly {
-			storageHandler.CompactIndexes(true)
-			storageHandler.CompactAll(storageFileTypeMeta, int(optGcDeadSkip))
-			storageHandler.CompactAll(storageFileTypeData, int(optGcDeadSkip))
+			storageHandler.store.CompactIndexes(true)
+			storageHandler.store.CompactAll(storagedb.StorageFileTypeMeta, int(optGcDeadSkip))
+			storageHandler.store.CompactAll(storagedb.StorageFileTypeData, int(optGcDeadSkip))
 		}
 		core.Log(core.LogInfo, "Garbage collection completed in %.1f minutes", time.Since(start).Minutes())
 	})
 
-	cmd.Command("recover", "[start file number] [end file number]", func() {
+	cmd.Command("recover", "[start-file-number] [end-file-number]", func() {
 		startfile := int32(0)
 		endfile := int32(-1)
 		if len(cmd.Args) > 2 {
 			i, err := strconv.ParseInt(cmd.Args[2], 0, 32)
-			abortOn(err)
+			core.AbortOnError(err)
 			startfile = int32(i)
 			core.Log(core.LogInfo, "Starting from file #%d (%08x)", startfile, startfile)
 		}
 		if len(cmd.Args) > 3 {
 			i, err := strconv.ParseInt(cmd.Args[3], 0, 32)
-			abortOn(err)
+			core.AbortOnError(err)
 			endfile = int32(i)
 			core.Log(core.LogInfo, "Stopping after file #%d (%08x)", endfile, endfile)
 		}
 
 		if lock, err := lockfile.Lock(filepath.Join(datDirectory, "hashbox.lck")); err != nil {
-			abort("%v", err)
+			core.Abort("%v", err)
 		} else {
 			defer lock.Unlock()
 		}
 
+		accountHandler = NewAccountHandler()
+		defer accountHandler.Close()
+		storageHandler = NewStorageHandler()
+		defer storageHandler.Close()
+
 		start := time.Now()
 
 		core.Log(core.LogInfo, "Checking storage file headers")
-		if storageHandler.CheckStorageFiles() > 0 {
+		if storageHandler.store.CheckStorageFiles() > 0 {
 			core.Log(core.LogError, "Recovery will not continue because errors where raised.")
 			returnValue = 1
 			return
 		}
 
 		core.Log(core.LogInfo, "Scanning data files")
-		repairCount := storageHandler.RecoverData(startfile, endfile)
+		repairCount := storageHandler.store.RecoverData(startfile, endfile)
 
 		if repairCount > 0 {
 			core.Log(core.LogInfo, "Performed %d repairs. Please run a verify.", repairCount)
@@ -462,11 +452,16 @@ func run() (returnValue int) {
 	cmd.Command("verify", "", func() {
 		if !optReadOnly {
 			if lock, err := lockfile.Lock(filepath.Join(datDirectory, "hashbox.lck")); err != nil {
-				abort("%v", err)
+				core.Abort("%v", err)
 			} else {
 				defer lock.Unlock()
 			}
 		}
+
+		accountHandler = NewAccountHandler()
+		defer accountHandler.Close()
+		storageHandler = NewStorageHandler()
+		defer storageHandler.Close()
 
 		errorCount := 0
 		start := time.Now()
@@ -474,19 +469,17 @@ func run() (returnValue int) {
 		core.Log(core.LogInfo, "Verifying dataset storage")
 
 		verifiedBlocks := make(map[core.Byte128]bool) // Keep track of verifiedBlocks blocks
-		rootlist := []BlockSource{}
-		if optReadOnly {
-			rootlist = accountHandler.CollectAllRootBlocks(false)
-		} else {
-			rootlist = accountHandler.RebuildAccountFiles()
+		if !optReadOnly {
+			accountHandler.RebuildAccountFiles()
 		}
+		rootlist := accountHandler.CollectAllRootBlocks(false)
 
 		for i, r := range rootlist {
 			tag := fmt.Sprintf("%s.%s.%x", r.AccountName, r.DatasetName, r.StateID[:])
 			core.Log(core.LogDebug, "Verify data referenced by %s", tag)
-			if err := storageHandler.CheckBlockTree(r.BlockID, verifiedBlocks, optVerifyContent, optReadOnly); err != nil {
+			if err := storageHandler.store.CheckBlockTree(r.BlockID, verifiedBlocks, optVerifyContent, optReadOnly); err != nil {
 				if optReadOnly {
-					abort("%v", err)
+					core.Abort("%v", err)
 				} else {
 					core.Log(core.LogWarning, "Dataset %s is marked as invalid: %v", tag, err)
 					accountHandler.InvalidateDatasetState(r.AccountNameH, r.DatasetName, r.StateID)
@@ -499,7 +492,7 @@ func run() (returnValue int) {
 		}
 
 		core.Log(core.LogInfo, "Verifying unreferenced index entries")
-		storageHandler.CheckIndexes(verifiedBlocks, optVerifyContent, optReadOnly)
+		storageHandler.store.CheckIndexes(verifiedBlocks, optVerifyContent, optReadOnly)
 
 		core.Log(core.LogInfo, "Verify completed in %.1f minutes", time.Since(start).Minutes())
 		if errorCount > 0 {
@@ -507,8 +500,7 @@ func run() (returnValue int) {
 		}
 	})
 
-	err = cmd.Parse()
-	abortOn(err)
+	core.AbortOnError(cmd.Parse())
 
 	fmt.Println(core.MemoryStats())
 
